requirements.txt
audioop-lts==0.2.2
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pydub==0.25.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
